SKALE-1772 Rename ledger.py
