+++ skale/wallets/__init__.py
@@ -1,4 +1,4 @@
 # flake8: noqa: E402
 
-from skale.wallets.ledger import LedgerWallet
+from skale.wallets.ledger_wallet import LedgerWallet
 from skale.wallets.web3_wallet import Web3Wallet
